Add files via upload
Changes made: *Add some new pages and images *Tried and currently failing to make some improvements to said images. Namely, I can't seem to get the images on the "Newer" page to work, but the images on the "Flyers" page do just fine and I'm not sure why. I notice that the "Flyers" images scale and the "Newer" images don't. Why that is, I have no clue.